Made database functnailtiy work
Created a working database that corrrectly stores the user input data and allows users to see the data in the backend database made with sql. Also added multiple visual tweaks to the overall look of the PWA including the bowling batting home and session summary page mainly the visual aspects.